+++ src/game/Strategic/Map_Screen_Interface.cc
@@ -1797,7 +1797,7 @@ static void DisplayUserDefineHelpTextRegions(FASTHELPREGION* pRegion)
 
 	{ SGPVSurface::Lock l(FRAME_BUFFER);
 		SetClippingRegionAndImageWidth(l.Pitch(), 0, 0, SCREEN_WIDTH, SCREEN_HEIGHT);
-		UINT16* const pDestBuf = l.Buffer<UINT16>();
+		UINT32* const pDestBuf = l.Buffer<UINT32>();
 		RectangleDraw(TRUE, iX + 1, iY + 1, iX + iW - 1, iY + iH - 1, RGB(65, 57, 15), pDestBuf);
 		RectangleDraw(TRUE, iX,     iY,     iX + iW - 2, iY + iH - 2, RGB(227, 198, 88), pDestBuf);
 	}
