+++ src/game/Strategic/Map_Screen_Interface_Map.cc
@@ -2545,7 +2545,7 @@ static void BlitMineText(UINT8 const mine_idx, const SGPSector& sMap)
 static void BlitTownGridMarkers(void)
 {
 	SGPVSurface::Lock l(guiSAVEBUFFER);
-	UINT16* const buf   = l.Buffer<UINT16>();
+	UINT32* const buf   = l.Buffer<UINT32>();
 	UINT32  const pitch = l.Pitch();
 
 	ClipBlitsToMapViewRegionForRectangleAndABit(pitch);
@@ -2615,7 +2615,7 @@ static void BlitMineGridMarkers(void)
 		w = MAP_GRID_X;
 		h = MAP_GRID_Y;
 
-		RectangleDraw(TRUE, x, y - 1, x + w, y + h - 1, color, l.Buffer<UINT16>());
+		RectangleDraw(TRUE, x, y - 1, x + w, y + h - 1, color, l.Buffer<UINT32>());
 	}
 
 	RestoreClipRegionToFullScreenForRectangle(pitch);
@@ -3746,7 +3746,7 @@ static void BlitSAMGridMarkers()
 		w = MAP_GRID_X;
 		h = MAP_GRID_Y;
 
-		RectangleDraw(TRUE, x, y - 1, x + w, y + h - 1, color, l.Buffer<UINT16>());
+		RectangleDraw(TRUE, x, y - 1, x + w, y + h - 1, color, l.Buffer<UINT32>());
 	}
 
 	RestoreClipRegionToFullScreenForRectangle(uiDestPitchBYTES);
