+++ src/game/Utils/Message.cc
@@ -107,7 +107,7 @@ static void BlitString(VIDEO_OVERLAY* pBlitter)
 
 	SetFontAttributes(pBlitter->uiFontID, pBlitter->ubFontFore, DEFAULT_SHADOW, pBlitter->ubFontBack);
 	SGPVSurface::Lock l(pBlitter->uiDestBuff);
-	MPrintBuffer(l.Buffer<UINT16>(), l.Pitch(), pBlitter->sX, pBlitter->sY, pBlitter->codepoints);
+	MPrintBuffer(l.Buffer<UINT32>(), l.Pitch(), pBlitter->sX, pBlitter->sY, pBlitter->codepoints);
 }
