+++ src/sgp/Line.h
@@ -20,8 +20,8 @@ void SetClippingRegionAndImageWidth(int iImageWidth, int iClipStartX, int iClipS
 //	Don't send fClip==TRUE to LineDraw if you don't have to. So if you know
 //  that your line will be within the region you want it to be in, set
 //	fClip == FALSE.
-void PixelDraw(BOOLEAN fClip, INT32 xp, INT32 yp, UINT32 sColor, UINT16* pScreen);
-void LineDraw(BOOLEAN fClip, int XStart, int YStart, int XEnd, int YEnd, UINT32 Color, UINT16* ScreenPtr);
-void RectangleDraw(const BOOLEAN fClip, const int XStart, const int YStart, const int XEnd, const int YEnd, const UINT32 Color, UINT16 *ScreenPtr);
+void PixelDraw(BOOLEAN fClip, INT32 xp, INT32 yp, UINT32 sColor, UINT32* pScreen);
+void LineDraw(BOOLEAN fClip, int XStart, int YStart, int XEnd, int YEnd, UINT32 Color, UINT32* ScreenPtr);
+void RectangleDraw(const BOOLEAN fClip, const int XStart, const int YStart, const int XEnd, const int YEnd, const UINT32 Color, UINT32 *ScreenPtr);
 
 #endif
